Change PeerSessionInfos log message to debug level
We are no longer including the peer client information in each BlockChainSync log message (since it bloats the logs). However, this information can be useful for debugging purposes and the initial peering could have happened long in the past, so promote the PeerSessionInfos log message (note that each PeerSessionInfo includes the clientVersion string) from trace to debug verbosity so that it's easier to have this information logged periodically for debugging purposes.